Fix issue with d3 js tree for tool recommendations
Fixes d3 JS tree for tool recommendation mentioned in https://github.com/galaxyproject/galaxy/issues/15723

PR against 23.0 on Galaxy main is open (https://github.com/galaxyproject/galaxy/pull/15852), but it may take some time to be merged. Can this PR be merged so that we can test it on the EU? All the lining tests are passing. Only one test fails but that's not related to this PR I think.

ping @bgruening @ElectronicBlueberry  thanks a lot!