overwrite upload paramfile after updating with incoming
the paramfile is created by `galaxy.tools.actions.upload_common.create_paramfile()`
and stored in `database/tmp/upload_params_TMPNAME` with mode 600 (it's
created with tmpfile, i.e. the mode is independent of the used umask).

`galaxy.jobs.MinimalJobWrapper.__prepare_upload_paramfile()` copies this
file to the jow working dir.

Without the change `ToolEvaluator.build_param_dict()` sets the paramfile
to the file in the jow working dir, but it is overwritten immediately
by the value in `incoming`.