New version of interactivetool_wallace.xml (#217)
* Update interactivetool_wallace.xml

New version to have the last Wallace app and also because old version was crashing apparenlty. HEre I just initialize the last version but not reintegrate all the work done on the previous one to interact with Galaxy through ie-helper to import and export data from the app. Will see this in the future

* Update interactivetool_wallace.xml command