Update interactivetool_qgis3_34.xml try to make outputs in Galaxy history
Seems that this version of QGIS is not exporting output datasets (from the GQGIS output folder) in Galaxy history when stopping the QGIS instance. Searching why, it appears to me that here, compare to the previous version of QGIS Galaxy tool, we are using /galaxy/output as folder instead of "just" /output. Do you think this can be the reason why, or the new QGIS image is not using the same PATH ?